object-detection/requirements.txt
numpy==1.19.5
opencv_python==4.1.2.30
pillow
PyQt5==5.15.6
PyYAML==6.0
